+++ packages/gov-ui-kit/src/components/index.ts
@@ -10,4 +10,5 @@ export * from './progress';
 export * from './spinner';
 export * from './switch';
 export * from './tag';
+export * from './textAreas';
 export * from './toggles';