+++ packages/gov-ui-kit/src/components/input/inputContainer/index.ts
@@ -1,2 +1,7 @@
 export { InputContainer } from './inputContainer';
-export type { IInputComponentProps, IInputContainerProps, InputVariant } from './inputContainer.api';
+export type {
+    IInputComponentProps,
+    IInputContainerProps,
+    InputComponentElement,
+    InputVariant,
+} from './inputContainer.api';