+++ packages/gov-ui-kit/tailwind.config.js
@@ -130,6 +130,7 @@ module.exports = {
             '2xl': 'var(--ods-border-rounded-2xl)',
             '3xl': 'var(--ods-border-rounded-3xl)',
             full: 'var(--ods-border-rounded-full)',
+            none: 'var(--ods-border-rounded-none)',
         },
         boxShadow: {
             'neutral-sm': 'var(--ods-shadow-neutral-sm)',
@@ -222,6 +223,7 @@ module.exports = {
         },
     },
     plugins: [
+        require('@tailwindcss/typography'),
         require('tailwindcss/plugin')(({ addVariant }) => {
             addVariant('search-cancel', '&::-webkit-search-cancel-button');
             addVariant('calendar-icon', ['&::-webkit-calendar-picker-indicator', '&::-webkit-inner-spin-button']);