+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,14 +9,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `Link`, `InputNumber` and `InputTime` components
--   Implement 'Addon' element for `InputText` component
+-   Implement `Link`, `InputNumber` `InputTime` and `TextAreaRichText` components
+-   Implement Addon element for `InputText` component
 -   Handle size property on `Progress` component
+-   `border-none` Tailwind CSS utility class
 
 ### Changed
 
 -   Update minor and patch versions of dependencies
 -   Update `husky` to v9
+-   Add `wrapperClassName` property to `InputContainer` component to customise the input wrapper
+-   Update `InputContainer` props to accept any HTML div property and support textarea elements
 
 ## [1.0.9] - 2024-01-23